requirements.txt
tensorflow>=2.10
numpy>=1.23
pillow>=9.0
streamlit>=1.20
streamlit-drawable-canvas>=0.9
matplotlib>=3.7
opencv-python>=4.7